# docs/swagger.yaml
definitions:
...
  models.Paper:
    properties:
      abstrak:
        type: string
      author:
        type: string
      file_paper:
        type: string
      id:
        type: integer
      judul:
        type: string
      link:
        type: string
      tanggal_terbit:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /papers:
    get:
      description: Retrieves all papers from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Paper'
            type: array
      summary: Get All Papers
      tags:
      - Papers
    post:
      consumes:
      - multipart/form-data
      description: Uploads a paper along with its file and saves them to the database.
      parameters:
      - description: Judul paper
        in: formData
        name: judul
        required: true
        type: string
      - description: Abstrak paper
        in: formData
        name: abstrak
        required: true
        type: string
      - description: Link paper
        in: formData
        name: link
        required: true
        type: string
      - description: File paper
        in: formData
        name: file_paper
        required: true
        type: file
      - description: Author paper
        in: formData
        name: author
        required: true
        type: string
      - description: Tanggal terbit paper
        in: formData
        name: tanggal_terbit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Paper'
      summary: Upload Paper with File
      tags:
      - Papers
  /papers/{id}:
    delete:
      description: Deletes a paper and its file from the database and storage.
      parameters:
      - description: Paper ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paper deleted successfully
          schema:
            type: string
      summary: Delete Paper
      tags:
      - Papers
    put:
      consumes:
      - multipart/form-data
      description: Edits a paper including the ability to replace its file.
      parameters:
      - description: Paper ID
        in: path
        name: id
        required: true
        type: string
      - description: Judul paper
        in: formData
        name: judul
        required: true
        type: string
      - description: Abstrak paper
        in: formData
        name: abstrak
        required: true
        type: string
      - description: Link paper
        in: formData
        name: link
        required: true
        type: string
      - description: File paper (optional)
        in: formData
        name: file_paper
        type: file
      - description: Author paper
        in: formData
        name: author
        required: true
        type: string
      - description: Tanggal terbit paper
        in: formData
        name: tanggal_terbit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Paper'
      summary: Edit Paper
      tags:
      - Papers
  /papers/file/{id}:
    get:
      description: Retrieves the file of a paper by its ID.
      parameters:
      - description: Paper ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Get Paper File
      tags:
      - Papers
swagger: "2.0"